pyproject.toml
[project]
name = "periodic_database_jobs"
description = "Periodic run database jobs"
license = {text = "MIT" }
version = "1.0"
dependencies = [
	"dacite",
	"postgresql_access",
	"keyrings.efile",
    "psycopg[binary]>=3.1",
    "PyYAML>=6.0",
    "dateparser>=1.2",
]
requires-python= ">= 3.10"
readme = "README.md"
authors = [
    {name = "Gerard"},
    {email = "gweatherby@uchc.edu"}
]

[tool.setuptools]
package-dir = {"" = "src" }  # Tells setuptools that the root package is in the "src" directory

[tool.setuptools.packages.find]
where = ["src"]  # Specify the "src" directory to look for packages


[project.scripts]
periodic_database_jobs = "periodic_database_jobs.main:main"

[project.optional-dependencies] 
# test = ['pytest']
ipython = ['ipython']

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "nmrbox"
url = "https://apt.nmrbox.org/pip"

